Change flags to use int64_t so that additional flags can be added Cleanup commented code Fix docs Allow string commtype argument to is_comm_installed Allow language to be passed to is_comm_installed but issue deprecation warning

+++ cpp/include/communicators/RMQComm.hpp
@@ -121,7 +121,7 @@ namespace YggInterface {
        */
       explicit RMQComm(const std::string name,
 		       const utils::Address& address,
-		       const DIRECTION direction=NONE, int flgs=0,
+		       const DIRECTION direction=NONE, uint64_t flgs=0,
 		       const COMM_TYPE commtype=RMQ_COMM);
       ADD_CONSTRUCTORS(RMQ)
